Don't publish PRs

+++ tools/publish_release.sh
@@ -1,26 +1,27 @@
-echo -e "Starting to update packages on GitHub"
+if [ "$TRAVIS_PULL_REQUEST" == "false" ]
+then
+    echo -e "Starting to update packages on GitHub..."
 
-# Go to the root of the repo
-cd ..
+    # Go to the root of the repo
+    cd ..
 
-cd ..
+    cd ..
 
-git clone https://${GH_TOKEN}@github.com/CatBlock/catblock-nightlies.git
+    git clone https://${GH_TOKEN}@github.com/CatBlock/catblock-nightlies.git
 
-rm -rf /builds
-rm catblock-edge.zip
-rm catblock-firefox.xpi
-rm catblock-chrome.zip
-rm catblock-opera.zip
+    rm -rf /builds
 
-rsync -r catblock/builds/* catblock-nightlies
+    rsync -r catblock/builds/* catblock-nightlies
 
-cd catblock-nightlies/
+    cd catblock-nightlies/
 
-git add -f .
+    git add -f .
 
-git commit -am "Updated packages by Travis $TRAVIS_BUILD_NUMBER"
+    git commit -am "Updated packages by Travis #$TRAVIS_BUILD_NUMBER"
 
-git push
+    git push
 
-echo -e "Done updating packages on GitHub"
+    echo -e "Done updating packages on GitHub."
+else
+    echo -e "Pull requests are not being shipped to CatBlock Nightlies repository."
+fi
